src/07_trees/leetcode/medium/102_BinaryTreeLevelOrderTraversal.ts
// Given the root of a binary tree, return the level order traversal of its nodes' values.
// (i.e., from left to right, level by level).

// Example 1:
// Input: root = [3,9,20,null,null,15,7]
// Output: [[3],[9,20],[15,7]]

// Example 2:
// Input: root = [1]
// Output: [[1]]

// Example 3:
// Input: root = []
// Output: []

// Constraints:

// The number of nodes in the tree is in the range [0, 2000].
// -1000 <= Node.val <= 1000

// Medium: https://leetcode.com/problems/binary-tree-level-order-traversal/description/

export class TreeNode {
    val: number;
    left: TreeNode | null;
    right: TreeNode | null;
    constructor(val?: number, left?: TreeNode | null, right?: TreeNode | null) {
        this.val = val === undefined ? 0 : val;
        this.left = left === undefined ? null : left;
        this.right = right === undefined ? null : right;
    }
}

export default function BinaryTreeLevelOrderTraversal(root: TreeNode | null) {
    if (!root) return [];
    let result = [];
    let queue = [root];

    while (queue.length) {
        let levelSize = queue.length;
        let currentLevel = [];

        for (let i = 0; i < levelSize; i++) {
            let node = queue.shift();
            currentLevel.push(node!.val);

            if (node!.left) queue.push(node!.left);
            if (node!.right) queue.push(node!.right);
        }

        result.push(currentLevel);
    }

    return result;
}
